refactor(adfluxbackend): 调整广告分类接口位置，优化用户注册参数校验 - 将广告分类接口从AdminController移动到AdvertiserController - 更新广告审核状态枚举，新增待审核状态 - 配置Jackson将Long类型序列化为String - 优化用户注册参数校验逻辑，增加用户名非空判断
