+++ src/main/java/com/usst/adfluxbackend/config/JsonConfig.java
@@ -1,0 +1,27 @@
+package com.usst.adfluxbackend.config;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import com.fasterxml.jackson.databind.ser.std.ToStringSerializer;
+import org.springframework.boot.autoconfigure.jackson.Jackson2ObjectMapperBuilderCustomizer;
+import org.springframework.boot.jackson.JsonComponent;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+
+/**
+ * Spring MVC json配置
+ */
+@Configuration
+public class JsonConfig {
+
+    @Bean
+    public Jackson2ObjectMapperBuilderCustomizer jackson2ObjectMapperBuilderCustomizer() {
+        return builder -> {
+            // 将 Long 类型序列化为 String
+            builder.serializerByType(Long.class, ToStringSerializer.instance);
+            // 将 long 基本类型序列化为 String
+            builder.serializerByType(Long.TYPE, ToStringSerializer.instance);
+        };
+    }
+}
